// client/src/app/_models/product.ts
export class Product {
    _id: string;
    productId: string;
    productName: string;
    productDesc: string;
    geoType: string;
    genType: string;
    wearType: string;
    categoryType: string;
    size: 
    [{
        small: string;
        medium: string;
        large: string;
        xlarge: string;
        
    }];
    rating: number;
    popularity: string;
    totalAvailability: string;
    price: number;
    discount: string;
    frontImage: string;
    backImage: string;
}